# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CoinswapApi V1 REST API Documentation",
        "description": "This is the documentation for the V1 of the rest API",
        "contact": {
            "email": "rafiumoshoodolakunle@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/coins": {
            "get": {
                "tags": [
                    "Coins"
                ],
                "summary": "All supported coin",
                "operationId": "all supported coin",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/coins/{coin}/show": {
            "get": {
                "tags": [
                    "Coins"
                ],
                "summary": "Show coin",
                "operationId": "show coin",
                "parameters": [
                    {
                        "name": "coin",
                        "in": "path",
                        "description": "e.g btc, eth, bch",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}